// src/app/shared/services/claim/pre-authorization/entities/provider.dto.ts
/***
 * Service Provider Information
 */
export interface ProviderDto {
    providerAuthId: number;
    facilityName: string;
    facilityRegisterNumber: string;
    physicianName: string;
    physicianDirectPhone: string;
    physicianOtherPhone: string;
    physicianAddress: string;
    physicianEmail: string;
    physicianLicense: string;
    physicianSpecialty: string;
    physicianOtherProfessionals: string;
    city: string;
    country: string;
    contactName: string;
    directPhone: string;
    otherPhone: string;
    email: string;
    department: string;
}